Fixes issue #294: deploy - startpage.
This is done by checking if the browser has anything from the player in local
storage. If this is the case, a splash screen is shown once (and only once).

Use the parameter lastVersion to test this feature like so:

  http://<our site>/lyt-<latest version>/#bookshelf?lastVersion=0.0.2

Using the address as above may confuse the browser, requiring a reload.

Besides the above, any tag with the class 'lyt-version' will have its inner
html replaced with the current version of the player upon startup.