Sort bookmarks reverse chronologically
This commit fixes bookmark sorting, which before was in respect to the
bookmark title which made no sense what-so-ever. This fix will not sort
the array every time a bookmark is added, but instead just figure out
the newly added bookmarks relative position, and then add it.

Note, that bookmarks that are already placed in books will *not* get
sorted due to the algorithm above.