Fix issue where iOS player would stall if first click was on play button
If the first click issued by the user (on an iOS device) is on the play
button, both the playbackrate tests and the silentplay command will be
started at the same time. For some reason this makes the silentplay
command hang and in turn the whole player. We fix this by stopping the
click event from bubbling in the LYT.player.playClickHook