Wrap controls in position:fixed header
iOS do for whatever reason not respect heights on <iframe> elements.
They render all the content in the iframe, and the height of it is
non-negotiable. Because of that scrolling on iOS failed, but with a
position:fixed header, this is no longer an issue.

A new issue is, that word highligthing/scrolling segments into view
doesn't seem to work