Stop using getSectionBySegment replaced by Segment.beginSection
When a SMIL document is loaded, all segments are parsed. We now mark all
the segments in the SMIL that is the beginning of a new section. We
utilize this in the renderSegment function, because it's way more
performant than iterating the SMIL DOM for each and every segment we
want to render

Known bug: When starting in the middle of a section (from a bookmark)
the playing section title isn't set