Extended integrated tests:
 - Feature tests for QUnit.
 - Fixtures that handles login, logout, loading and playing books.
 - A number of utility functions that extends jQuery mobile.
 - A Chain utility for QUnit to clean up the syntax when making heavy use of
   deferreds in tests.

To prepare the tests, get a copy of fixtures.json (not checked in because it
contains sensitive information) and link (or copy) it into build/test.

To run the tests, go to the page #test and click "Kør tests" or just run
QUnit.start() from the console.

Not ready for integration into master as it is now, because we need a build
flag that leaves out the test definitions when building for live.