Support for multivolume DTB
On request from MTM we now support multivolume DTBs - that is, books
containing several ncc files and a single discinfo.html file
orchestrating it all.

The main shift has been to rely on "localUris" instead of the actual URL
given in a document. When loading a new file, it's always important to
know in which directory it lies.