Implement option for custom "skinning"
This commit implements build option "-s [dir]" for choosing a specific skin.
This skin directory *MUST* contain a "package.json" file, which in turn
*MUST* have a property ".config" which is the path of a
configuration file.

If you want to add or overwrite pages in the player, make a
"directories" object in the package.json, and have a ".pages" property
point to the path of a directory containing the html pages.

If a page is named the same, as a page in the original project (e.g.
"profile.html" or "about.html"), the original file will be overwritten.