Use simple object persistence if localstorage isn't available
This is a possible fix for MS-1333, that a lot of users have
experienced. Since private mode Safari (and maybe others) exposes
localStorage, but doesn't persist anything in it, LYT was tricked into
believing that it actually saved the users credentials.

Now if localStorage doesn't persist, we just save data in an object in a
closure.